First cut at documenting custom CAs in rotation
This new file documents the additional steps that a customer must go
through during an Ops Manager certificate rotation in order to handle
any customer-provided CAs that they may have installed.

This file has not been linked to from anywhere else in the
documentation, as we're uncertain of the correct place to make such a
link from.

[#177731776] Document how to handle `generated: false` CAs in CredHub during Ops Manager cert rotation

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>